Add a Condition to the openModal Function in App
Add a condition to the openModal function to handle the isEditing variable. When the user opens the modal for editing a cookie, the isEditing variable is set to true. However, if the user closes the modal without saving or exits the modal, the condition will check for this and reset the value of isEditing to false. This ensures that when the user later presses the add cookie button, any previous cookie data will not appear in the form, preventing any potential bugs.# Please enter the commit message for your changes. Lines starting