Welcome Text, Query Search, Popular Movies Fetch
This commit implements the following changes:

Adds a welcome text to the header of the application with a transition effect.
Implements a function that takes a query from the user and returns the movie(s) that match the query. The results are then saved in the movie list.
Implements another function that fetches a list of popular movies. The fetched list is then saved in the movie list.
Calls the function to fetch popular movies in the onMounted lifecycle hook. This ensures that when the app is loaded, all the movies will be displayed.